moved index into journalist_app/main.py
