+++ securedrop/journalist_templates/base.html
@@ -33,7 +33,7 @@
     <div class="content">
       {% block header %}
       <div id="header">
-        <a href="{{ url_for('index') }}"><img src="/static/i/{{ header_image }}" class="logo small" alt="SecureDrop" width="250px"></a>
+        <a href="{{ url_for('main.index') }}"><img src="/static/i/{{ header_image }}" class="logo small" alt="SecureDrop" width="250px"></a>
         {% include 'locales.html' %}
         {% if use_custom_header_image %}
         <div class="powered">
